feat: Include 'sendAndReceiveData' function
Add the 'sendAndReceiveData' function to send data to server3 for generating commentary and receiving a response. This function sends an HTTP POST request with the filtered event data to the specified server URL and waits for a response. The response is then logged, and further processing can be added as needed.